* feat(handlers.coreDatas.py): 增加数据处理控件，分离自run.py * feat(handlers.coreLink.py): 增加友链处理控件，分离自run.py * feat(handlers.coreSettings.py): 增加增加全局设置控件初始值和debug模式自检 * feat(component.getTime.py): 增加时间信息处理组件 * feat(component.ohter.py): 增加其他组件，备用 * feat(request_data.request.py): 打印调整 * feat(run.py): 重构主程序结构，减肥 * feat(runOld.py): 备份主程序 * feat(settings.py): 增加一些变量 * feat(run.py): 增加debug模式时使用不同的精益云 * style(*): None * todo(settings.py): 将请求器部分超时设置迁移过来 * todo(coreDatas.py): 数据处理leancloud对象传递分离扔不完全或可能存在问题。
